+++ hookee/plugins/response_json.py
@@ -1,11 +1,10 @@
-import click
 from flask import jsonify
 
 from hookee import pluginmanager
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.7"
+__version__ = "0.1.0"
 
 plugin_type = pluginmanager.RESPONSE_PLUGIN
 print_util = None
@@ -20,11 +19,5 @@ def setup(cli_manager):
 def run(request, response):
     if not response:
         response = jsonify({"Response": "Ok"})
-
-    click.secho("Status Code: {}".format(response.status_code), fg="magenta")
-    if response.headers:
-        print_util.print_dict("Headers", dict(response.headers), fg="magenta")
-    if response.data:
-        print_util.print_dict("Body", response.get_json(), fg="magenta")
 
     return response